requirements.txt
confusable-homoglyphs==3.1.1
dj-database-url==0.5.0
Django==1.11.5
django-appconf==1.0.2
django-bootstrap4==0.0.6
django-imagekit==4.0.2
django-map-widgets==0.1.9
django-registration==2.4.1
gunicorn==19.8.1
pilkit==2.0
Pillow==5.1.0
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
